move reducers into src/functions
