+++ src/ctx/index.ts
@@ -2,8 +2,5 @@ export * from './DeckContext';
 export * from './GameContext';
 export * from './SettingContext';
 export * from './deckDefaults';
-export * from './deckReducer';
 export * from './gameDefaults';
-export * from './gameReducer';
 export * from './settingDefaults';
-export * from './settingReducer';
