+++ src/functions/index.ts
@@ -1,3 +1,6 @@
 export * from './cardKeys';
+export * from './deckReducer';
 export * from './fisherYates';
+export * from './gameReducer';
 export * from './getRandomIndex';
+export * from './settingReducer';
