+++ src/functions/settingReducer.ts
