+++ src/types/MessageBarDefinition.ts
@@ -1,5 +1,8 @@
 import { MessageBarType } from "@fluentui/react"
 
+/**
+ * Definition of a message bar
+ */
 export type MessageBarDefinition = {
   type: MessageBarType;
   text: string;
